+++ rust/operator-binary/src/config/mod.rs
@@ -1,5 +1,5 @@
 pub mod client_protocol;
 pub mod fault_tolerant_execution;
 pub mod jvm;
-pub mod openlineage;
+pub mod lineage;
 pub mod s3;