+++ rust/operator-binary/src/trino_controller.rs
@@ -371,7 +371,7 @@ mod tests {
             trino_opa_config,
             resolved_fte_config,
             resolved_client_protocol_config,
-            resolved_open_lineage_config: None,
+            resolved_lineage_config: None,
         };
 
         let operator_env = OperatorEnvironmentOptions {
@@ -603,7 +603,7 @@ mod tests {
             trino_opa_config: None,
             resolved_fte_config: None,
             resolved_client_protocol_config: None,
-            resolved_open_lineage_config: None,
+            resolved_lineage_config: None,
         };
         let operator_env = OperatorEnvironmentOptions {
             operator_namespace: "stackable-operators".to_string(),